requirements.txt
websocket-client
yfinance
ccxt
pandas_ta
requests
pandas
